+++ packages/rum/src/domain/record/observers/inputObserver.spec.ts
@@ -3,9 +3,9 @@ import { createNewEvent } from '@datadog/browser-core/test'
 import { PRIVACY_ATTR_NAME, PRIVACY_ATTR_VALUE_MASK_USER_INPUT } from '../../../constants'
 import { serializeDocument, SerializationContextStatus } from '../serialization'
 import { createElementsScrollPositions } from '../elementsScrollPositions'
-import { DEFAULT_CONFIGURATION, DEFAULT_SHADOW_ROOT_CONTROLLER } from '../../../../test'
 import type { InputCallback } from './inputObserver'
 import { initInputObserver } from './inputObserver'
+import { DEFAULT_CONFIGURATION, DEFAULT_SHADOW_ROOT_CONTROLLER } from './observers.specHelper'
 
 describe('initInputObserver', () => {
   let stopInputObserver: () => void
