+++ packages/rum/src/domain/record/observers/observers.specHelper.ts
@@ -1,0 +1,12 @@
+import { noop } from '@datadog/browser-core'
+import type { RumConfiguration } from '@datadog/browser-rum-core'
+import type { ShadowRootsController } from '../shadowRootsController'
+import { NodePrivacyLevel } from '../../../constants'
+
+export const DEFAULT_SHADOW_ROOT_CONTROLLER: ShadowRootsController = {
+  flush: noop,
+  stop: noop,
+  addShadowRoot: noop,
+  removeShadowRoot: noop,
+}
+export const DEFAULT_CONFIGURATION = { defaultPrivacyLevel: NodePrivacyLevel.ALLOW } as RumConfiguration
